agent/requirements.txt
fastapi==0.112.1
uvicorn==0.30.6
google-genai==0.3.0
numpy>=1.24.3
torch>=2.2.0
transformers>=4.36.0
sentence-transformers>=2.2.2
raganything
lightrag
